requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-admin-honeypot-updated-2021==1.2.0
django-admin-thumbnails==0.2.9
django-ipware==7.0.1
idna==3.8
pillow==10.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-ipware==3.0.0
razorpay==1.4.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2